perf(consensus): Remove reactor rlocks on Consensus (#3211)
PR with @ebuchman !!!

Remove (most) of the reactor RLock's on consensus! (Remaining one is in
the gossip routine, should be easy to fix, but as a separate PR)

The consensus reactor on ingesting messages takes RLock's on Consensus
mutex, preventing us from adding things to the queue. (And therefore
blocking behavior)

This can cause the peer message queue to be blocked. Now it won't be
blocked because we update the round state directly from the cs state
update routine (via event bus) when:
- A vote is added
- A block part is added
- We enter a new consensus step. (Receiving a full block triggers this
for us, we will enter prevote or precommit round)

This shouldn't change reactor message validity, because Reactor.Receive
could always view a cs.RoundState that will be old when the packet
reaches the cs.HandleMsg queue. Every update to consensus' roundstate
pushes the update into the reactor's view, so we avoid locks.

I don't think any new tests are required!

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Ethan Buchman <ethan@coinculture.info>
Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>