test(e2e): reverting PR3211 since it is making e2e nightly to fail (#3335)
This PR reverts #3211 since it is making the e2e nightly to fail in
reproducible ways. This PR was identified as the reason for the recent
e2e nightly failure on `main` and doing a bi-sect test of all recent
commits it was determined that this PR introduced a behavior that makes
the tests to fail and indicate a bug or unknown behavior has been
introduced.

Even though we are reverting this logic for now, we'd be happy to
consider it in the future again once more tests are performed and we can
ensure it passes all tests.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [ ] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec