refactor(mempool)!: remove PreUpdate from Mempool interface and move its logic to Lock (#3363)
Complements #3361 

This PR partially reverts the backport of
https://github.com/cometbft/cometbft/pull/3314 into the recently
released v0.37.7. With this change the Mempool interface is the same as
in previous versions.

The reason is that we do not want to break the public API. We still keep
in the code the feature that
https://github.com/cometbft/cometbft/pull/3314 introduced by moving it
inside the existing Lock method. We also keep the RecheckFull bool field
that we added to ErrMempoolIsFull.

<!--

Please add a reference to the issue that this PR addresses and indicate
which
files are most critical to review. If it fully addresses a particular
issue,
please include "Closes #XXX" (where "XXX" is the issue number).

If this PR is non-trivial/large/complex, please ensure that you have
either
created an issue that the team's had a chance to respond to, or had some
discussion with the team prior to submitting substantial pull requests.
The team
can be reached via GitHub Discussions or the Cosmos Network Discord
server in
the #cometbft channel. GitHub Discussions is preferred over Discord as
it
allows us to keep track of conversations topically.
https://github.com/cometbft/cometbft/discussions

If the work in this PR is not aligned with the team's current
priorities, please
be advised that it may take some time before it is merged - especially
if it has
not yet been discussed with the team.

See the project board for the team's current priorities:
https://github.com/orgs/cometbft/projects/1

-->

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [ ] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec