test(e2e): add option to the 'runner logs' command to output logs separately (#3353)
When conducting end-to-end testing, it can be difficult to effectively
analyze or visualize the output of docker logs when all the containers
are merged into a single log file.

To address this issue, I have introduced a new parameter (`--split`) to
the runner `logs` command. When this parameter is used, the command will
output the log for each container separately.

Having separate logs makes it much easier to troubleshoot individual
containers, particularly when the log output is extensive and detailed.
This feature boosts efficiency by simplifying the process of identifying
the root cause of failures.

Usage:
```sh
 ./build/runner logs -f ./networks/simple.toml --split
```
#### PR checklist

- [ ] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec