fix(e2e): test case `TestBlock_Range` is flaky because of pruning (#3389)
This PR is an attempt to solve -- once and for all -- the flakiness in
`e2e` test case `TestBlock_Range`. The sequence of commits in this PR
shows the repro steps, the fix, and the final contents of the patch.
It's not necessary to review it commit by commit, unless the reviewer
wants to check the steps taken to reproduce the flakiness.

* Commits 1 to 3: Infrastructure + temporary modifications to have the
flakiness of `TestBlock_Range` show at every run. Also, some further
fixes in the checks of `TestBlock_Range` are improved.
* Commit 4: Reduce `e2e` tests to be run so that the GitHub action
doesn't time out. This commit reproduces the flakiness: click the ❌ next
to this commit below.
* Commit 5: The fix. The test now passes, even with all the infra to
repro in place. Check the ✅ next to the commit below.
* Commit 6: revert commit 4.
* Commit 7: Remove all the repro infra.
* Commit 8: Fix a typo that makes `e2e` CI tests fail.

All other commits: addressing issues in workflows and comments.

---

#### PR checklist

- [x] Tests written/updated
- ~[ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)~
- ~[ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments~
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Andy Nogueira <me@andynogueira.dev>