refactor(mempool): Add `Iterator` to replace `TxsFront` and `TxsWaitChan` methods (#3459)
Closes #3303 

This PR adds:
- the `Entry` interface, implemented by the existing `mempoolTx` data
type,
- the `Iterator` interface, with one method `WaitNextCh() <-chan Entry`.

On the routine that sends transactions to a peer, we can now use an
`Iterator` to select the next mempool entry to send, instead of the
methods `TxsFront` and `TxsWaitChan` in `CListMempool`. These two
methods leak implementation details about the CList and are now marked
as deprecated. Now all the logic needed to wait for and pick the next
available entry is implemented in `CListMempool.WaitNextCh`.

---

#### PR checklist

- [X] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>