refactor(config/node)!: Remove genesisHash from config (#3595)
Closes #3547 .

This PR removes the `GenesisHash` from the config file because this
field is not meant to be updated by the operator. It was only introduced
to avoid breaking an API and used to pass the genesis hash provided via
cli to the node constructor.


As this feature is going into an unreleased version, we decided to break
the API of `DefaultNewNode`. We maintain the signature of `NewNode` as
this is probably used by the community more and introduce
`NewNodeWithCliParams`.

`NewNodeWithCliParams` allows for passing different parameters from the
cli to the node.

The PR also introduces a new type `CliParams`. At the moment this struct
contains just the `GenesisHash` but it was done this way so that we can
in the future extend it with new parameters without breaking APIs.

---------

Co-authored-by: Daniel <daniel.cason@informal.systems>