feat(indexer): update psql indexer to have configurable table names (#3594)
Closes #3593

- Updates the psql indexer to accept custom table names.
- Introduces optional arguments to the `NewEventSink(...)` constructor
as to retain backwards compatibility. This includes all four table names
as well as a `*sql.DB` reference for test or other usage.
- The `NewEventSink(...)` is modified to add a passed in `*sql.DB`
instance to the struct and if that is not set, then create a new
database connection and set within the event sink.
- Renames the default table names constants with a prefix `default` to
make clear where they're used.
- Stores configured table names within the `EventSink` struct instead of
references the global default variables.
- Updates tests to use constructor to exercise that code path. Manually
constructing the struct does not set the defaults within the struct.
- Adds four new options to the `.toml` config which get passed to the
`NewEventSink` should any be set (not an empty string).
- `make test` passes

---

#### PR checklist

- [X] Tests written/updated
- [X] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [X] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [X] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Daniel <daniel.cason@informal.systems>
Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>