test(e2e): add support for testing different key types, including BLS (#3513)
While working on another crypto to be added to CometBFT, I realised we
are not testing other key types in our nightlies, in particular BLS
isn't tested.

This PR:
* adds support for BLS in the runner, generator, and `e2e` tests
* adds support for BLS, secp256k1, and ed25519 key types in the `e2e`
generator (for the nightlies)
* completes support for changing the key type at testnet level. Some
support was existing but it wasn't complete (so it didn't work)

There is a TODO in the BLS crypto side. I added a temporary workaround
and contacted the Zondax folks to include a small patch in
`github.com/cosmos/crypto`. If they agree to merge the patch,
`GenPrivKeyFromSecret` will become trivial.

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- ~[ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments~
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>