docs: ADR-117 `PrivValidator` should implement `cosmos.CryptoProvider` (#3400)
<!--

Please add a reference to the issue that this PR addresses and indicate
which
files are most critical to review. If it fully addresses a particular
issue,
please include "Closes #XXX" (where "XXX" is the issue number).

If this PR is non-trivial/large/complex, please ensure that you have
either
created an issue that the team's had a chance to respond to, or had some
discussion with the team prior to submitting substantial pull requests.
The team
can be reached via GitHub Discussions or the Cosmos Network Discord
server in
the #cometbft channel. GitHub Discussions is preferred over Discord as
it
allows us to keep track of conversations topically.
https://github.com/cometbft/cometbft/discussions

If the work in this PR is not aligned with the team's current
priorities, please
be advised that it may take some time before it is merged - especially
if it has
not yet been discussed with the team.

See the project board for the team's current priorities:
https://github.com/orgs/cometbft/projects/1

-->

This PR introduces **ADR-117** which implements the `CryptoProvider`
[ADR](https://github.com/cosmos/crypto/blob/main/docs/architecture/adr-001-crypto-provider.md)
("aka pluggable cryptography")

The ADR in this repo is only a shortened version of the original one
posted [here ](https://github.com/cosmos/cosmos-sdk/pull/18824)
containing only the relevant concepts and code for cometBFT.

For the full text explaining all concepts please check this
[file](https://github.com/cosmos/crypto/blob/main/docs/architecture/adr-001-crypto-provider.md)
(links are included in ADR-117)


---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [X] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [X] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec