build(crypto/secp256k1): Update `btcec/v2` to v2.3.4
Running a `go get` should consistently build this project as per
the README. However, the latest patch release of `btcec/v2` is not
a proper Semantic Versioning patch. It removes an error return value
from `ecdsa.SignCompact` function (btcsuite/btcd@e5d15fd). This is
functionally a no-op, as the underlying function was hardcoded to
always return `nil` as error. However, this is still a breaking change
(compiler/build error) for all consumers of this function, as they
historically expected two return values, but now only get one.