test(mempool): Allow custom log levels on reactor tests (#3719)
Also disable the P2P log messages on the switches, which are not
relevant for mempool.

In particular, this PR changes the log level of
`TestReactor_MaxTxBytes`, which is very verbose and it currently could
take 12s with the default "debug" level. With "info" level it takes
300ms.

For reviewers, the PR can be reviewed commit by commit.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>