build(crypto/secp256k1): Update `btcec/v2` to v2.3.4 (#3728)
Running a `go get` should consistently build this project as per the
README. However, the latest patch release of `btcec/v2` is not a proper
Semantic Versioning patch. It removes an error return value from
`ecdsa.SignCompact` function (btcsuite/btcd@e5d15fd). This is
functionally a no-op, as the underlying function was hardcoded to always
return `nil` as error. However, this is still a breaking change
(compiler/build error) for all consumers of this function, as they
historically expected two return values, but now only get one.

Ref: btcsuite/btcd#2211
Ref: #3531
Ref: #3536

---

#### PR checklist

- [ ] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>