feat(lanes/e2e): let loader define the lane of each tx (#3825)
In e2e, transactions generated with `runner load` have a constant prefix
`"a="`, which does not work with how lanes are currently assigned. This
PR lets `runner load` to set the lane of each tx that it generates, and
it includes it in the tx's `Payload`. Then on `CheckTx`, the app parses
the payload to obtain the lane.

Note that e2e's `TestApp_Tx` generates transaction values that are not
`Payload`s, so for these cases, `CheckTx` should not fail if it cannot
parse the tx, it must just assign the 0 lane.

The loader assigns lanes picked randomly from the list of lanes defined
in the app. This PR adds to the e2e manifest an option
`load_lane_weights`, which is used by the loader to select the lane
proportionally to its weight.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: Sergio Mena <sergio@informal.systems>