test(mempool/lanes): Use app with lanes in tests + do not test ordering (#3987)
Depends on #3978 and #3980 (then we should merge feature branch into
target branch).

This PR removes from kvstore the option `useLanes`, so it's not possible
to disable lanes once the app is created. All tests use kvstore with
lanes as default, except those that initialise the app with
`New...ApplicationWithoutLanes`.

Also, some reactor tests were failing when checking that txs were
received in order. We only need to check that the txs are in the
mempool; order is tested elsewhere.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>