feat(e2e): simplify the relation between manifest and testnet (#3964)
Refactors the `e2e` testbed to remove some unnecessary duplication
between `Manifest` and `Testnet`, and `ManifestNode` and `Node`. Fields
that are identical in the two types are removed from the latter, which
includes (extends) the former. Fields that had the same name but
different types or semantics in the two versions, were renamed in the
manifest.

With this change, to add a new field to the configuration of a testnet
or a node, it is enough to update the `Manifest` or `ManifestNode`. The
new field must also be added to `Testnet` or `Node` if it needs to be
somehow parsed (see the examples of the `Seeds` and `PersistentPeers`
fields).

---

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Andy Nogueira <me@andynogueira.dev>