perf: add metrics in mempool, block executor; and remove `BlockProcessingTime` (#3891)
Contributes to #3864

These new metrics are needed to monitor some of the scenarios described
in #3864, where the bottleneck is not the sender's p2p rate limiter.

Removing `BlockProcessingTime`, as the metric is confusing (it's in ms,
in contrast to other metrics expressing duration). Besides, the metric
`MethodTimingSeconds` (for label `FinalizeBlock`) is measuring the same.

---

#### PR checklist

- ~[ ] Tests written/updated~
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>