feat(consensus): proposer locks part set until all parts are added (#4009)
Addresses #1742.

When a validator is the round proposer, it produces a `Proposal` and a
number of block `Part`s that it then delivers to itself
([code](https://github.com/cometbft/cometbft/blob/ec2f3e2bc7eebe1dba69ef2a681c20fd973f332f/internal/consensus/state.go#L1258-L1271)).
When a `Proposal` is accepted, the validator initializes a
`types.PartSet` instance from the proposed `BlockID`
([code](https://github.com/cometbft/cometbft/blob/4952be833d1b1254f79f0e03b862f3adfba4e64f/internal/consensus/state.go#L2091-L2093)).
Then, when each of the block `Part`s are delivered, they are added to
this part set. Once a `Part` is added to the `cs.ProposalBlockParts`
part set, it becomes available to the consensus reactor, that sends them
to peers.

The issue reported in #1742 is that we add the block parts in order to
`cs.ProposalBlockParts`, and one by one, even when the we are the
proposer. When we are the proposer we know all the block parts and could
produce `cs.ProposalBlockParts` immediately, instead of processing all
`Part`s, one by one, as they were received from the network. This would
have as a consequence that the consensus reactor will only start sending
block `Part`s to peers once all of them are available. Since we randomly
choose the next part to send to a peer, we are less likely to send to
all peers the same part. That is what currently happens, as reported in
#1742.

The solution proposed here is relatively simple and non-intrusive. To
the `types.PartSet` type we add a boolean `locked` value. Once the
proposer receives its own `Proposal`, it creates and _locks_ its
`cs.ProposalBlockParts`, which is unlocked when all the expected parts
are received and processed.

The consensus reactor, in turn, only send blocks `Part`s to peers when
`cs.ProposalBlockParts` is unlocked. When the node is not the proposed
of the round, it is always unlocked. When the node is the current round
proposer, it is unlocked once all the parts it should contain are added,
likely fixing the issue reported in #1742.

---

#### PR checklist

- [x] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Sergio Mena <sergio@informal.systems>