fix(e2e): viper errors with invalid `config` and `genesis` manifest keys (#4016)
Addresses comments
https://github.com/cometbft/cometbft/pull/3969#pullrequestreview-2277655582
and
https://github.com/cometbft/cometbft/pull/3967#pullrequestreview-2277635023
on previous PRs.

The solution is in fact simple, as it is possible to tell viper to
return an error if one of the configured entries (key/value pairs) were
not applied to the destination structure, because they don't exist or
are invalid.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Andy Nogueira <me@andynogueira.dev>
(cherry picked from commit bff1667f0a9e48916baffe2430352f01ee8b6612)

# Conflicts:
#	test/e2e/pkg/testnet.go
#	test/e2e/runner/setup.go