test(lanes/e2e): Fix load weights (#4137)
`load_lane_weights` was not updated with the changes in #4092.

Also in this PR:
- Change default list of lanes and priorities in e2e app. Now it has 5
default lanes, each with 1000 txs capacity.
- Do not use "default" as default lane ID. Just use one random id from
the list of lanes. If we didn't include "default" in the list of lanes
in the manifest, nodes were starting and then crashing.
- Change the `lane` field in `Payload` message: from `payload.Lane` to
`string` (for lane id). The priority is not needed, so I also removed
`payload.Lane` from the proto file.