refactor(e2e): change `*Node` to string in keys of some maps (#4141)
Partially addresses #2458

Change the `*Node` as keys of some maps in `e2e` and replace them by the
node's name.

Having pointers as keys of a map brings about subtleties (how comparison
is done) that we don't need. Code is much clearer with node names, even
if it takes a bit longer to execute some loops (we need to look up the
nodes). But this is test code, so clarity and robustness comes before
performance of the testing code.

I decided to tackle this refactor in an different PR, rather than
piggybacking it to #2283

---

#### PR checklist

- [x] Tests written/updated
- ~[ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)~
- ~[ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments~

---------

Co-authored-by: Andy Nogueira <me@andynogueira.dev>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Daniel <daniel.cason@informal.systems>