test(mempool): enhanced `TestReactorNoBroadcastToSender` (#4127)
The test should be able now to catch the error introduced by
https://github.com/cometbft/cometbft/pull/3657.

----

To see how it would break, please checkout the branch
`cason/test-nobroadcastsender`, then:

```
$ cd mempool
$ go test -v -run TestReactorNoBroadcastToSender
```

And wait it to fail, after 2 minutes. You are going to see the busy-loop
around, logs entries starting with `Skipping tx`, forever for the same
transaction.

Important: the test unit only works in the traditional mempool design,
without lanes. We should consider adding a variant of it that should
also pass when lanes are considered.

---

#### PR checklist

- [x] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>