docs(Node): Expand docs of `GenesisDocProvider` (#4232)
### Context
`GenesisDocProvider` is a function that users can implement and provide
when creating a new `Node`. The function allows users to decide how the
`GenesisDoc` should be provisioned to the chain. When called, it returns
a
[`ChecksummedGenesisDoc`](https://github.com/cometbft/cometbft/blob/87a8c8948bf9926aff37f344b4120163f12cb4ad/node/setup.go#L46),
an object that stores both the `GenesisDoc` and its SHA256 checksum.

Our code does not verify the internal consistency of the
`ChecksummedGenesisDoc`. Specifically, it does not check whether the
`Sha256Checksum` field is actually the hash of the `GenesisDoc` field it
contains. We rely on the `GenesisDocProvider` to ensure that
`SHA256(GenesisDoc) == Sha256Checksum`. However, this critical
assumption isn't explicitly documented in the `GenesisDocProvider` docs.

### This Change
This PR updates the documentation to clarify that it's the implementer's
responsibility to verify the internal consistency between the
`GenesisDoc` and its SHA256 checksum.

---

#### PR checklist

~- [ ] Tests written/updated~
~- [] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)~
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments

(cherry picked from commit 7b048da65d2df434f0e563c297f9ab1a4b40e729)