The `ConfigureRPC` method of `Node` now treats the `Environment` object that it returns as a singleton.
The purpose of these changes is to set the `Node`'s pointer to the genesis
to `nil` after initialization, allowing the garbage collector to deallocate
the genesis struct from memory. In future work, we plan to remove this
pointer entirely.

Currently, different parts of the code call the `ConfigureRPC()` method to
create an `Environment` object for their own purposes. However, this is
wasteful because each `Environment` instance stores a copy of the genesis
file: either as a single pointer or as a collection of chunks.
Therefore, creating multiple, independent `Environment` objects results in
multiple copies of the genesis being loaded into memory. For large genesis
files, this can lead to large memory usage. By treating `Environment` as a
singleton within `ConfigureRPC()`, we prevent multiple copies of the
genesis from being loaded into memory simultaneously.