Set node ptr to genesis to nil after startup is complete.
The `Node` does not need to access the genesis after starting up.
If the genesis is small (<=16MB) then the `Environment` object serving the RPC
APIs will retaing a pointer to the genesis at runtime.
If the genesis is big (>16MB) then the `Environment` object will split the
genesis into multiple chunks and keep them in memory.