`Environment.InitGenesisChunks` makes a local copy of chunks.
It seems that the `base64.StdEncoding.EncodeToString` function does not make a
full copy of its argument. Therefore creating each `chunk` as a direct sub-slice
of `data` was preventing the GC from deallocating the `data` slice after the
function returned (you could still see it as allocated in pprof graphs).

Creating each `chunk` as a copy of the sub-slice of `data` solves the issue.