feat(p2p)!: extract MConnection from Peer
Contributes to #4302

**It's now the job of a `Transport` implementation to emit a multiplexed
connection.**

Previously, `MConnection` was created inside `Peer`, and `Transport` was
responsible for returning a raw `net.Conn`. With the upcoming QUIC
transport, this is no longer acceptable. Now, TCP transport returns a
multiplexed connection (`MConnection`). QUIC transport, likewise, will
be emitting multiplexed connection too. The difference is, of course, in
case of TCP it's out-of-protocol multiplexing.

Note I've made an internal "queue" in `MConnection`, which stores
packets and the associated "channelID" (streamID). So, `MConnection` has
a `readLoop`, which pushes incoming packets onto this queue
(`map[streamID]chan []byte` actually). The packets then get read in
`Peer#streamReadLoop` (1 per stream), which calls `Stream#Read`. So even
though there's a single TCP connection, there's an abstraction of
streams of data.