Added new functions to start RPC HTTP server.
These functions optionally accept a list of callbacks to execute at shutdown.
We use this feature to delete the genesis file's chunks from disk when the
server shuts down by passing a function that handles the deletion as a
parameter.
However, we could pass any other function we wish to execute at this stage.