Improved `mkChunksDir`.
The function now first checks for and deletes any existing chunks directory.
While a Node's graceful shutdown procedure deletes the genesis file chunks and
their enclosing directory, if the Node (or its RPC HTTP server) crashes, it's
possible that the genesis file chunks remain on disk and are present upon restart.

This is not necessarily a problem, but if the genesis file changes for any reason,
its chunks will also differ. Therefore, to be safe, we re-create the chunks from
scratch at startup.