perf(types/vote_set): Remove BlockID.Key usage for equality in AddVote (#4324)
Partially addresses #2848 by removing the calls to `Key` in `AddVote`
when doing equality checks.

Since `Equals checks compare all of the fields in the BlockID and nested
BlockPartSet, this is functionally the same and avoids the allocations
that Key incurs.


This includes a benchmark that shows the reduction in memory
allocations.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Alessandro Sforzin <alessandro@informal.systems>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>
Co-authored-by: Ryan Ramas <orionpulse61@gmail.com>
Co-authored-by: wesl-ee <w@wesl.ee>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Hernán Vanzetto <15466498+hvanz@users.noreply.github.com>