`BenchmarkLoadValidators` no longer panics.
The code was creating a new state `Store` without specifying the `DBKeyLayout`
in the `StoreOptions`. Consequently, the `NewStore` function invoked
`setDBKeyLayout` with an empty string, resulting in an incorrect key layout for
the database. This caused the `LoadValidators` call in the benchmark to fail
because the key did not match any entries in the database.